% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requireR.R
\name{requireR}
\alias{requireR}
\title{Loads given files and modularizes the given function}
\usage{
requireR(...)
}
\arguments{
\item{...}{A set of filenames, and/or ending with the modularizable function.}
}
\description{
Loads given files and modularizes the given function
}
\examples{
##hello.R
#requireR(function() {
#  "hello"
#})

##world.R
#requireR(function() {
#  "world"
#})

#requireR(
#  "hello.R",
#   "world.R",
#   function(hello, world) {
#     paste(hello, world)
#   })
}

